requirements.txt
fastapi
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
python-multipart==0.0.6
jinja2>=2.11.2
swagger-ui-bundle>=0.0.8
alembic==1.12.1
python-dateutil>=2.8.0
pytest
httpx
pydantic[email]